Refactor lint() to makeLinter()
